Refactor getDshotSensorData
