+++ src/main/drivers/dshot.h
@@ -29,6 +29,7 @@
 #include "drivers/motor_types.h"
 // TODO: move bitbang as implementation detail of dshot (i.e. dshot should be the interface)
 #include "drivers/dshot_bitbang.h"
+#include "sensors/esc_sensor.h"
 
 #define DSHOT_MIN_THROTTLE              (48)
 #define DSHOT_MAX_THROTTLE              (2047)
@@ -151,5 +152,5 @@ void dshotCleanTelemetryData(void);
 float erpmToRpm(uint32_t erpm);
 
 int16_t getDshotTelemetryMotorInvalidPercent(uint8_t motorIndex);
-
+bool getDshotSensorData(escSensorData_t *dest, int motorIndex);
 void validateAndfixMotorOutputReordering(uint8_t *array, const unsigned size);
